% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots accidents for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number for plot}

\item{year}{Accident year for plot}
}
\value{
Nothing returned, produces a side effect of a plot of given US state and location of accident(s) for given year
}
\description{
Plots accidents for a given state and year
}
\details{
This function will also not plot accidents where their longitudes are more than 900 or latitudes are more than 90.
}
\examples{
\dontrun{
fars_map_state(50, 2012)
}
}
