% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data from single file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path of CSV containing FARS data}
}
\value{
A data frame containing data loaded from indicated filename.
}
\description{
Read FARS data from single file
}
\examples{
\dontrun{
fars_read("dummy.csv")
}
}
