% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise FARS data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{integer vector containing the years of FARS data to read}
}
\value{
A data frame containing the year, month and count of data in each corresponding year/month
}
\description{
Summarise FARS data by year and month
}
\examples{
\donttest{
test <- fars_summarize_years(2013)
}
}
