% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generates filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year to be indicated in filename}
}
\value{
A character string for the filename containing the year provided.
}
\description{
Generates filename
}
\examples{
make_filename(2012)
}
